Move custom mp4 code from KTAC
