+++ src/libvgmstream.vcxproj
@@ -184,6 +184,7 @@
     <ClCompile Include="coding\coding_utils.c" />
     <ClCompile Include="coding\ffmpeg_decoder.c" />
     <ClCompile Include="coding\ffmpeg_decoder_utils.c" />
+    <ClCompile Include="coding\ffmpeg_decoder_custom_mp4.c" />
     <ClCompile Include="coding\ffmpeg_decoder_custom_opus.c" />
     <ClCompile Include="coding\lsf_decoder.c" />
     <ClCompile Include="coding\mp4_aac_decoder.c" />
