+++ src/libvgmstream.vcxproj.filters
@@ -1819,6 +1819,9 @@
     <ClCompile Include="coding\ffmpeg_decoder_utils.c">
       <Filter>coding\Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="coding\ffmpeg_decoder_custom_mp4.c">
+      <Filter>coding\Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="coding\ffmpeg_decoder_custom_opus.c">
       <Filter>coding\Source Files</Filter>
     </ClCompile>
